refactor: изменение запросов в классе ReviewDbStorage, style: форматирование кода.
