+++ src/main/java/ru/yandex/practicum/controller/EventController.java
@@ -2,14 +2,12 @@ package ru.yandex.practicum.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ru.yandex.practicum.service.EventService;
 import ru.yandex.practicum.model.event.Event;
+import ru.yandex.practicum.service.EventService;
 
 import java.util.List;
