+++ src/main/java/ru/yandex/practicum/model/film/Director.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.model.film;
 
-import lombok.Data;
 import lombok.AllArgsConstructor;
+import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import javax.validation.constraints.NotBlank;
