+++ src/main/java/ru/yandex/practicum/model/film/Film.java
@@ -3,13 +3,9 @@ package ru.yandex.practicum.model.film;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NonNull;
-import ru.yandex.practicum.db.FilmReleaseDate;
+import ru.yandex.practicum.validation.FilmReleaseDate;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.PastOrPresent;
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.Size;
+import javax.validation.constraints.*;
 import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.List;
