+++ src/main/java/ru/yandex/practicum/service/FilmService.java
@@ -12,13 +12,9 @@ import ru.yandex.practicum.model.film.MPA;
 import ru.yandex.practicum.storage.DirectorStorage;
 import ru.yandex.practicum.storage.FilmStorage;
 import ru.yandex.practicum.storage.UserStorage;
-import ru.yandex.practicum.db.FilmValidator;
+import ru.yandex.practicum.validation.FilmValidator;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 @Slf4j
 @RequiredArgsConstructor
