+++ src/main/java/ru/yandex/practicum/storage/db/UserDbStorage.java
@@ -13,12 +13,7 @@ import ru.yandex.practicum.model.user.FriendConnection;
 import ru.yandex.practicum.model.user.User;
 import ru.yandex.practicum.storage.UserStorage;
 
-import java.sql.Date;
-import java.sql.ResultSet;
-import java.sql.Statement;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
+import java.sql.*;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.LinkedList;
