+++ src/main/java/ru/yandex/practicum/validation/FilmReleaseDate.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.db;
+package ru.yandex.practicum.validation;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
