+++ src/main/java/ru/yandex/practicum/validation/ReviewValidator.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.db;
+package ru.yandex.practicum.validation;
 
 import ru.yandex.practicum.exception.ValidationException;
 import ru.yandex.practicum.model.film.Review;
